Moved collision point methods onto their own file.
