+++ lib/CMakeLists.txt
@@ -23,7 +23,8 @@ add_library(react_controller include/react_controller/controllerNLP.h
                              src/react_controller/controllerNLP.cpp
                              src/react_controller/ctrlThread.cpp
                              src/react_controller/mathUtils.cpp
-                             src/react_controller/baxterChain.cpp)
+                             src/react_controller/baxterChain.cpp
+                             src/react_controller/collisionPoints.cpp)
 
 ## Add cmake target dependencies of the library
 ## as an example, code may need to be generated before libraries
